+++ LICENSE-MIT
@@ -1,4 +1,4 @@
-Copyright (c) 2018 Tokio Authors
+Copyright (c) 2018 Tokio Contributors
 
 Permission is hereby granted, free of charge, to any
 person obtaining a copy of this software and associated
