+++ tokio-io/README.md
@@ -26,14 +26,13 @@ online at [https://tokio.rs](https://tokio.rs). The [API
 documentation](https://docs.rs/tokio-io) is also a great place to get started
 for the nitty-gritty.
 
-
 # License
 
 This project is licensed under either of
 
- * Apache License, Version 2.0, ([LICENSE-APACHE](LICENSE-APACHE) or
+ * Apache License, Version 2.0, ([LICENSE-APACHE](../LICENSE-APACHE) or
    http://www.apache.org/licenses/LICENSE-2.0)
- * MIT license ([LICENSE-MIT](LICENSE-MIT) or
+ * MIT license ([LICENSE-MIT](../LICENSE-MIT) or
    http://opensource.org/licenses/MIT)
 
 at your option.
