+++ tokio-reactor/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "tokio-reactor"
+
+# When releasing to crates.io:
+# - Update html_root_url.
+# - Update CHANGELOG.md.
+# - Create "v0.1.x" git tag.
+version = "0.1.0"
+authors = ["Carl Lerche <me@carllerche.com>"]
+license = "MIT/Apache-2.0"
+readme = "README.md"
+repository = "https://github.com/tokio-rs/tokio"
+homepage = "https://tokio.rs"
+documentation = "https://docs.rs/tokio-reactor/0.1"
+description = """
+Event loop that drives Tokio I/O resources.
+"""
+categories = ["asynchronous", "network-programming"]
+
+[dependencies]
+futures = "0.1.18"
+log = "0.4.1"
+mio = "0.6.13"
+slab = "0.4.0"
+tokio-executor = { version = "0.1", path = "../tokio-executor" }
+tokio-io = { version = "0.1", path = "../tokio-io" }
